requirements.txt
ultralytics>=8.3.97
streamlit>=1.32.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0 